pyproject.toml
[tool.poetry]
name = "streamdeck-magierdinge"
version = "0.1.0"
description = ""
authors = ["RincewindWizzard <git@magierdinge.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamdeck = "^0.9.5"
pillow = "^10.4.0"
cairosvg = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
